+++ src-tauri/src/game_dev_api/v1_handlers.rs
@@ -3,38 +3,13 @@ use axum::{
     extract::{Query, State},
     http::StatusCode,
     response::IntoResponse,
-    Json, Router,
+    Json,
 };
-use axum_macros::FromRef;
 use serde::{Deserialize, Serialize};
-use serde_json::{from_str, Value};
-use std::sync::Arc;
-use std::{option::Option, path::PathBuf};
-use tokio::{
-    sync::watch::Receiver,
-    sync::{Notify, RwLock},
-};
+use serde_json::Value;
+use std::option::Option;
 
-// TODO: rename to not be confused with the managed tauri app state
-#[derive(Clone, FromRef)]
-pub struct AppState {
-    pub api_state: ApiState,
-    pub game_state: GameStateShared,
-}
-
-#[derive(Clone)]
-pub struct ApiState {
-    pub database_path: String,
-}
-
-#[derive(Debug, Clone)]
-pub struct GameState {
-    pub id: Arc<RwLock<Option<u64>>>,
-    pub notifier: Arc<Notify>,
-    pub channel: Receiver<Option<u64>>,
-}
-
-pub type GameStateShared = Arc<GameState>;
+use super::{ApiState, GameStateShared};
 
 #[derive(Deserialize, Serialize)]
 pub struct LeaderboardPost {
@@ -57,9 +32,6 @@ pub async fn set_leaderboard_v1(
     State(game_state): State<GameStateShared>,
     Json(payload): Json<LeaderboardPost>,
 ) -> impl IntoResponse {
-    // TODO: Get game_id and user_id
-    println!("Setting Laaderboard data");
-    // let game_id = "1";
     let game_id = game_state.id.read().await.unwrap().to_string();
     drop(game_state);
     let user_id = payload.player_slot.to_string();
@@ -150,7 +122,6 @@ pub async fn set_save_data_v1(
     State(game_state): State<GameStateShared>,
     Json(payload): Json<SaveDataPost>,
 ) -> impl IntoResponse {
-    // let game_id = "0";x
     let game_id = game_state.id.read().await.unwrap().to_string();
     drop(game_state);
     let user_id = payload.player_slot.to_string();
